Refactor how card filtering works
