+++ pkg/rubbernecker/card.go
@@ -91,3 +91,40 @@ func (c Cards) Filter(s string) Cards {
 
 	return tmp
 }
+
+func (c Cards) FilterByStickerNames(
+	includeStickers []string,
+	excludeStickers []string,
+) Cards {
+	filteredCards := make(Cards, 0)
+
+	for _, card := range c {
+		shouldAdd := true
+
+		if len(includeStickers) > 0 {
+			shouldAdd = false
+			for _, sticker := range card.Stickers {
+				for _, includedStickerName := range includeStickers {
+					if sticker.Name == includedStickerName {
+						shouldAdd = true
+					}
+				}
+			}
+		} else if len(excludeStickers) > 0 {
+			shouldAdd = true
+			for _, sticker := range card.Stickers {
+				for _, excludedStickerName := range excludeStickers {
+					if sticker.Name == excludedStickerName {
+						shouldAdd = false
+					}
+				}
+			}
+		}
+
+		if shouldAdd {
+			filteredCards = append(filteredCards, card)
+		}
+	}
+
+	return filteredCards
+}
